GNORM-00: Refactoring for DRY maintainability.
