+++ src/CompilerInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Gnorm;
+
+interface CompilerInterface
+{
+    public function execute();
+}
