+++ src/GnormStaticSiteCompiler.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Gnorm;
+
+class GnormStaticSiteCompiler extends CompilerBase
+{
+
+}
